backend/requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
gitdb==4.0.11
GitPython==3.1.42
pytz==2024.1
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
